Spliti monarch_extension into a common shared library
Summary:
As we're implementing `MastAllocator` in `hyperactor_meta`, we cannot contaminate `monarch_extension` with its wrappers. Instead we would create a separate python rust extension that can be loaded separately.

However for the sake of consistency, we don't want to duplicate common constructs such as `AllocSpec`. For this purpose, this diff takes common stuff into a new rust_library target `hyperactor_extension` that can be used by `monarch_extention` as well as `hyperactor_meta`.

This is the first step towards adding `MastAlloc*` bindings.

Reviewed By: highker

Differential Revision: D73933989

fbshipit-source-id: a873d7b4536c213166d2e2af5391c608bf7e554e