Remove race condition for clean shutdown in terminal with `scribe_cat` subprocess.
Summary:
Before this diff, any error tha that happend while writing to the input buffer of a subprocess would trigger the scruba logger to print an error and switch to a noop variant of itself to reduce the changs of impacting a workload during a scribe outage etc.

I had missed that the flush calls also need to apply this treatment as `scribe_cat` may exit before the hyper actor process when killed by process group in the terminal.

Reviewed By: vidhyav

Differential Revision: D73954161

fbshipit-source-id: 876b3f1d1f07168a0e27fd49ffc30e0b7e6041ab