Use counter type that will produce delta encoding rather than cumlative sum
Summary:
Before this change we were using UpDownCounter, which defaults to a cumaltive sum encoding for export. This is because the value is expected to represent a value over a long period of time but that you don't have a single source of truth for like a gauge.

We want our counters to represent the sum of values that have been reported between two exports to scuba, so using Counter rather than UpDownCounter fixes this issue.

This delta encoding issue did not appear to impact any other metric types such as gauge or histogram.

Reviewed By: kaiyuan-li

Differential Revision: D74007114

fbshipit-source-id: 0ced5a15690ea3359f3408f508e5216b53c6b51b