Pass in dialer address when calling dial()
Summary:
This is to set us up for the next 2 diffs where we will remove source/destination semantics from SimAddress by passing in the source address when we dial an address

The important parts of this are in `channel/mod.rs` and `channel/sim.rs`. Everything else is just passing in `None` as the dialer for now. In the next diff we will make ProcActors and SystemActors pass in their listen addresses for this arg.

Full Context:

One of the biggest pain points with the simulator is that there is a concept of directionality in our addresses. This adds cognitive complexity as we have to reason about pointing the channel in the correct direction and making sure to always have the correct source. One of the kinks caused by this is for example, we introduced the capability for an actor to talk to another actor directly instead of going through the system after it learns its address, but the system would have to make sure to modify this address to use the right source each time instead of only being concerned with the address of the destination.

The reason why we need directionality for simulation is so that we can figure out how much latency should occur when a message is sent. The only piece that needs to know this is the transmitter, and for a transmitter, it does make sense semantically have a concept of directionality (I am sending to an address, from an address).

What we will do instead of source-to-destination addresses for sim channels, is we will store the source and destination information in the transmitter, and initialize these values when we dial a channel. dial() will now take in both the address we want to dial, as well as and optional address of the dialer. The most common use case is when a proc dials the system, it will give the address it is being served on as the dialer. The system will also need to bind it's own id to the address is it being served on so it can provide this when dialing other procs.

Reviewed By: moonli, pablorfb-meta

Differential Revision: D74854480

fbshipit-source-id: 3d522453b46029c68bbc2096d48f90b6741dcae6