[monarch] improve error behavior during pickle/unpickling
Today, if there is an error in unpickling the message content, it will kill the actor and generate a supervision event. We cannot respond on the port to generate an ActorError that the client can handle.

This is because the response Port itself we get from unpickling, so if there is an error during unpickling we don't have anywhere to send a response.

This diff moves the response port into the `PythonMessage` itself so that it can be unconditionally deserialized. I also had to remove `RankedPort` (since we can no longer rely on pickle to preserve that behavior), and make `rank_in_response` another first-class thing in `PythonMessage`.

I think this plus a few other changes should also be able to allow us to remove the mailbox serialization hack from `_pickle` and `_unpickle`, but I will address that in a forthcoming diff

Differential Revision: [D76171937](https://our.internmc.facebook.com/intern/diff/D76171937/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D76171937/)!

[ghstack-poisoned]