Avoid repeat error reports
If an op depends on two nodes that are failing at the time it is issued, then history will generate two failure messages for the sequence number. If there was a future attached to the sequence number, then the second failre will cause a KeyError in the invocation dicationary since the first instance already removed it.

This depends on a race condition to hit because the messages for the failure must already be present when the node failing is being added.

Differential Revision: [D76372510](https://our.internmc.facebook.com/intern/diff/D76372510/)

[ghstack-poisoned]