[hyperactor] py: ensure that proc mesh lifetimes are tied to actor mesh lifetimes
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/299

In Python, a ProcMesh may outlive its alloc. This is because the monitoring loop of the ProcMesh—which owns the alloc—is tied to its Python object, which may be dropped while there are extant actor mesh references.

---

The EventState owns the underlying alloc, because this is how we're driving all changes to the proc.

```
struct EventState {
    alloc: Box<dyn Alloc + Send + Sync>,
    supervision_events: PortReceiver<ActorSupervisionEvent>,
}
```

We transfer ownership to whomever is listening to events.

```
    /// An event stream of proc events. Each ProcMesh can produce only one such
    /// stream, returning None after the first call.
    pub fn events(&mut self) -> Option<ProcEvents> {
        self.event_state.take().map(|event_state| ProcEvents {
            event_state,
            ranks: self
                .ranks
                .iter()
                .enumerate()
                .map(|(rank, (proc_id, _))| (proc_id.clone(), rank))
                .collect(),
        })
    }
```

In PyProcMesh, we take over the event stream to perform monitoring. When PyProcMesh is then dropped, we abort the task, the closure, and thus the event stream. This causes the alloc to be dropped. It's the alloc that maintains the server.

---

We fix this by ensuring that the monitoring loop is kept alive as long as there are any Python proc mesh or actor mesh objects.

In the more medium term, this will be fixed by decoupling these lifetimes in the underlying APIs.
ghstack-source-id: 291273600
@exported-using-ghexport

Differential Revision: [D76873976](https://our.internmc.facebook.com/intern/diff/D76873976/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D76873976/)!