[6/n] Future gains result()/exception() and can be await more than once
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/289

Previously our ActorFuture would do weird stuff if get()/await was called more than once. Now it behaves like a normal future (records the result/exception and replay it on each get/await).

This also adds the result()/exception() method our monarch.common.Future had. This is for compat when the next PR flips the tensor engine over to ActorFuture.

For compat, we need to add timeout to result. The easiest way to achieve this is to just run it through the async path with a wait_for call. Otherwise we have to make sure _all_ blocking_impl paths take a timeout parameter, which is way more work.

This doesn't implement done() because it isn't easy: some of our "futures" require actual work to do before they complete so they cannot be polled for done like true futures, which are just a container.
ghstack-source-id: 291379180
@exported-using-ghexport

Differential Revision: [D76828364](https://our.internmc.facebook.com/intern/diff/D76828364/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D76828364/)!