[8/n] Get history ready for working with ports
This cleans up the history object in order to replace the message returns with direct sending to the return port.

In particular, it clarifies the lifetimes of Invocation objects an Ref s by replacing the hash tables with Arc links. Now invocations live as long as there is a Ref referring to them, or if it is still in flight. Status enum make sure we do not hold users longer than needed.

Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/278
ghstack-source-id: 291379176

Differential Revision: [D76649221](https://our.internmc.facebook.com/intern/diff/D76649221/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D76649221/)!