Make timed_test robust to deadlocked tokio runtimes
Make it so that the timeout is guaranteed to fire, even if the underlying runtime is completely blocked on some bogus work that never awaits.

This type of code should *never* be written; all native code that may deadlock should be run on an isolated thread via `spawn_blocking` or similar. However it's important to make our tests robust to it.

Differential Revision: [D77342496](https://our.internmc.facebook.com/intern/diff/D77342496/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D77342496/)!

[ghstack-poisoned]