[monarch] remove HYPERACTOR_MANAGED_SUBPROCESS
HYPERACTOR_MANAGED_SUBPROCESS is intended to ensure that child processes are promptly cleaned up after spawning.

It is implemented by setting up a SIGUSR1 handler on child processes, and calling PR_SET_DEATHSIG on parent to have the parent signal the children when it dies.

This can conflict with user signal handlers (for example, checkpointing uses SIGUSR1 to coordinate between the trainer and checkpointing processes).

Better to get rid of it. It's not necessary anymore, as we have a natural cleanup when the parent dies: the child is listening on a channel for `Allocator2Process` messages, and will error out when the channel is closed due to parent process death.

Differential Revision: [D77392241](https://our.internmc.facebook.com/intern/diff/D77392241/)

[ghstack-poisoned]