[hyperactor] general in-proc ActorRef -> ActorHandle downcasting
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/350

Provide a general purpose mechanism for downcasting any local (in-proc) ActorRef<A> to its corresponding ActorHandle<A>. This allows these references to be coordinated externally, but for the actors to communicate locally (including using handlers that aren't exported, e.g., because they may not contain serializable messages).

The intent is to use this in the tensor engine, to allow the controller to instruct actors to coordinate locally, passing objects like tensor references directly.

This required building a general instance registration mechanism, which we can use to replace the ledger in a follow-up change.
ghstack-source-id: 293003584

Differential Revision: [D77377382](https://our.internmc.facebook.com/intern/diff/D77377382/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D77377382/)!