[hyperactor] cleanup
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/356

In several places, we have a pattern of newtyping an Arc<State>, to provide a publicly accessible objects (e.g., Mailbox, Proc, etc.) that is cloneable.

This is just a small cleanup to standardize this pattern into

```
struct Outer {
  inner: Arc<Inner>,
}

...
```

Possibly we should also standardize on calling such objects "Handles", since they are not owned, but rather provide a handle to some other object.
ghstack-source-id: 293003582

Differential Revision: [D77395146](https://our.internmc.facebook.com/intern/diff/D77395146/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D77395146/)!