[monarch] delete extra runtime
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/363

Libraries should never create tokio runtimes; that should be up to the caller to manage.

Seems that we were doing this because pyo3 threads did not have the runtime context properly set. So just fix that and zap this runtime. This cuts the number of threads we spawn in the PythonActor by a lot.
ghstack-source-id: 293110124
@exported-using-ghexport

Differential Revision: [D77422927](https://our.internmc.facebook.com/intern/diff/D77422927/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D77422927/)!