[monarch] remove our custom signal handlers
These need to be written quite carefully as they need to be async-signal-safe (no log, no printf, no malloc). Let's just not even try, and instead rely on the standard fb implementation which has been worked over forever.

Normally `#[fbinit]` will do this on the main entry point, but because bootstrap_main is a Python entry point that we spawn via the ProcessAllocator we need to call `initFacebook` there as well.

Differential Revision: [D77565855](https://our.internmc.facebook.com/intern/diff/D77565855/)

[ghstack-poisoned]