[monarch] get rid of hyperactor runtime, attempt 2
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/406

Based on the discussion in D77422927, a new attempt:

I still want to avoid creating an extra runtime in hyperactor, which I think is a Pretty Bad thing to do.

So here, I have `hyperactor::initialize` take a runtime handle provided by the caller. This ensures the user has full control over the runtime, but that we still have a reference to a runtime handy in hyperactor when we need one from a sync context.
ghstack-source-id: 293872756
@exported-using-ghexport

Differential Revision: [D77632354](https://our.internmc.facebook.com/intern/diff/D77632354/)