[hyperactor] mesh: key streams off of sender only
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/397

Each stream was previously associated with `(sender, <shape of mesh>)`, the shape being a stand-in for the mesh itself. In practice, each proc is reachable through exactly one mesh, so this is unnecessary.

This change simplifies this to just the sender.

In the future, we'll likely want to support multiple sessions per sender (e.g., for persistent meshes), but that is a straightforward addition we can make as needed.
ghstack-source-id: 293881496
@exported-using-ghexport

Differential Revision: [D77609169](https://our.internmc.facebook.com/intern/diff/D77609169/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D77609169/)!