[monarch] 'this' -> 'cx' in handlers
Now that we have altered actor `handle` method to take a `Context`, let's also name the variables accordingly.

This was a mix of LLM edits and manual ones: I ended up trying Devmate, Claude, and Gemini.

Interestingly, this ended up being kryptonite for both Devmate and Claude. Gemini did a little better, but got itself into some weird loops too.

Here's the initial prompt:

```
 In hyperactor/src/actor.rs, we define a Handler trait:

`pub trait Handler<M>: Actor {     /// Handle the next M-typed message.     async fn handle(&mut self, this: &Context<Self>, message: M) -> Result<(), anyhow::Error>; }`

The reason the Context-typed argument is called 'this' is that the type used to be called 'Instance'.Can you rename all instances of this argument to 'cx' instead? And also all of the usages of course.Please do this for all of the rust files in this project.
```

After Gemini got a start, but got stuck, I tried Claude again with the following prompt:

```
This repository contains an in-progress change to change instances of "this: &hyperactor::Context<...>" to "cx: &hyperactor::Context<...>" and then update the usages      │
│   (this->cx). Can you finish the job?
```

It claimed it was done a few times, and I had to then cajole it into completing the task by showing grep output with evidence to the contrary. There were also a few instances of ... frankly strange hallucinations. For example, replacing `this` with `1`. Maybe the models are trying to keep us on our toes.

(It would be simpler if Devmate could just use the LSP to perform the renames. In that case, it would only need to identify the candidate sites, and instruct rust-analyzer to do the job.)

Differential Revision: [D77753150](https://our.internmc.facebook.com/intern/diff/D77753150/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D77753150/)!

[ghstack-poisoned]