[monarch] heartbeat in child processes to ensure cleanup
In D77392241, I removed the signal-based cleanup mechanism. The assumption there was that the `rx.recv()` in `bootstrap.rs` would throw an error if the other side hung up, and the error would bubble up and eventually abort the child process.

This assumption is wrong; `rx.recv()` has server-like semantics, not channel-like (which makes sense; there could be many senders, so any individual one disappearing should not abort the receiver!).

But as a result, we were not cleaning up properly if the parent process exited, child processes would just hang around forever. The test added in D77348271 *happened* to pass, because we sent `SIGKILL` to the parent process, triggering unclean shutdown which *will* cause a error on the receiver side . However, a *graceful* shutdown (e.g. from an uncaught Python exception) will not.

A quick solution is to add a simple heartbeat task and kill the process if it fails.

Differential Revision: [D77802426](https://our.internmc.facebook.com/intern/diff/D77802426/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D77802426/)!

[ghstack-poisoned]