[monarch] initial commit of tensor_engine submodule
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/458

Now that the actor layer is place, time to refactor the tensor engine following the same principles:

- Every symbol has at most two locations: a private location (which should be exactly the same as its location in the source tree), and optionally a public location in the user-facing API.
- We discourage end-users from depending on private locations via underscores, per python convention.

In particular, we violate principle 2 in many places in the tensor engine—all of our top-level files/modules are visible if you do `from monarch.foo import bar`, and we are sloppy about the distinction of public and private.

So this refactor will be a little bit painful. I will approach it in parts. This diff defines the `tensor_engine` submodule and moves `common` to `monarch.tensor_engine._common`.

Differential Revision: [D77847587](https://our.internmc.facebook.com/intern/diff/D77847587/)
ghstack-source-id: 294872880