[14/n] tensor engine {handle_cast, handle} -> handle
We unified the message handling logic in rust, so we can keep that unification for the python bindings. We keep the handle_cast variant, rename it handle.

We provide a definition for cast coordinates even when a message wasn't cast: the message is considered to be the  rank of a zero dimension shape.

Differential Revision: [D78061501](https://our.internmc.facebook.com/intern/diff/D78061501/)

ghstack-source-id: 295416054
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/491