Update on "[monarch] cut log spam"
The amount of spew when just trying to use monarch as a user was crazy lol.
Should fix https://github.com/pytorch-labs/monarch/issues/411.

This creates two layers:
- A file log layer
- A stderr log layer

They are configured separately depending on environment. The file one is controlled by RUST_LOG, and the stderr one is controlled by MONARCH_LOG.

For now, both are always used, even if redundant.

You can look at the proposed defaults on an environment basis, but for local/mast we have the same config: errors to stderr, info+ to file.

Differential Revision: [D78220912](https://our.internmc.facebook.com/intern/diff/D78220912/)

[ghstack-poisoned]