[monarch] cut log spam
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/524

The amount of spew when just trying to use monarch as a user was crazy lol.
Should fix https://github.com/pytorch-labs/monarch/issues/411.

This creates two layers:
- A file log layer
- A stderr log layer

They are configured separately depending on environment. The file one is controlled by RUST_LOG, and the stderr one is controlled by MONARCH_LOG.

For now, both are always used, even if redundant.

You can look at the proposed defaults on an environment basis, but for local/mast we have the same config: errors to stderr, info+ to file.
ghstack-source-id: 295862670
@exported-using-ghexport

Differential Revision: [D78220912](https://our.internmc.facebook.com/intern/diff/D78220912/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D78220912/)!