Update base for Update on "[15/n] Pass monarch tensors to actor endpoints, part 1"
This makes it possible to send a monarch tensor to an actor endpoint that is defiend over the same proc mesh as the tensor.

The send is done locally so the actor can do work with the tensors.

The stream actor in the tensor engine is sent a SendResultOfActorCall message, which it will forward to the actor, binding to the message the real local tensors that were passed as arguments.

The stream actor that owns the tensor waits for called the actor to finish since the actor 'owns' the tensor through the duration of the call.

Known limitation: this message to the actor can go out of order w.r.t to other messages sent from the owner of the tensor engine because the real message is being sent from the stream actor. The next PR will fix this limitation by sending _both_ the tensor engine and the actor a message at the same time. The actor will get a 'wait for SendResultOfActorCall' message, at which point it will stop processing any messages except for the SentResultOfActorCall message it is suppose to be waiting for. This way the correct order is preserved from the perspective of the tensor engine stream and the actor.

Differential Revision: [D78196701](https://our.internmc.facebook.com/intern/diff/D78196701/)

[ghstack-poisoned]