[18/n] tensor engine: Simplify CallFunctionError
The only case we every 'catch' is DependentError. Everything else is just a textural description. So this collapses those cases into just the two cases.

This retains the enum variants as function calls so that we can also reify them again if we need to catch them but having them obscures the places where DependentError needs to be handled differently.

Differential Revision: [D78363663](https://our.internmc.facebook.com/intern/diff/D78363663/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D78363663/)!

[ghstack-poisoned]