[20/n] tensor engine, more refactoring of dependent error handling
This PR introduces a SeqError type which represents the 'root' seq that failed, clearly delinated from dependent errors in the type system.

env objects are now changed to hold Arc<SeqError> with the lookup functions always returning a DependentError of the seq error.

This makes it clear that our error type does not nest dependent errors, and it makes it clear where  function invocation errors first get introduced.

I use this to introduce a single place that creates the Arc<SeqError>, reports the error to the controller, and sets it as the last error so that recordings know that error occurred.

This PR also removes partially finished work to propagate pipe creation errors to users of the pipe. We will not need pipes for much longer, and this functionality wasn't complete anyway.

Differential Revision: [D78683472](https://our.internmc.facebook.com/intern/diff/D78683472/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D78683472/)!

[ghstack-poisoned]