[23/n] tensor engine, implement rref on actor endpoints
Allows actor endpoints to take tensors as arguments and return them as monarch.Tensor objects.

Actor endpoints that do this must have a propagator function defined.

"cached" does not currently work because the caching machinery cannot make the actor call.

Introduces `as_endpoint()` which can take something that was not annotated on a class and turn it into an endpoint. This is useful when needing to define a per-actor propagation function for rref calls because there otherwise is nowhere to put it.

Differential Revision: [D78782587](https://our.internmc.facebook.com/intern/diff/D78782587/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D78782587/)!

[ghstack-poisoned]