[19/n] Refactor dependent error handling
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/619

Next diff will add rref calls for tensors so I am getting the call_function logic consistent so that a call function that is not calling a function can be implemented. This moves all the handling of error propagation for call_function into `try_define`, similar to python backend instead of handle_results.  It is possible other cases in stream.rs also should be converted over to use try_define as well to ensure consistent error propagation.
ghstack-source-id: 298043755
@exported-using-ghexport

Differential Revision: [D78683483](https://our.internmc.facebook.com/intern/diff/D78683483/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D78683483/)!